requirements.txt
# ── Core data science ────────────────────────────────────────────────
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
openpyxl==3.1.2
joblib==1.3.2

# ── Explainability ───────────────────────────────────────────────────
shap==0.44.1

# ── LangChain ecosystem (keep as-is) ─────────────────────────────────
langchain==0.2.16
langchain-core>=0.2.39,<0.3
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langchain-huggingface==0.0.3
langchain-groq==0.1.9
langgraph==0.2.28
langgraph-checkpoint==1.0.12

# ── LLM clients ──────────────────────────────────────────────────────
groq==0.9.0
httpx==0.27.2

fastparquet==2024.2.0

# ── Embeddings & vector store ────────────────────────────────────────

sentence-transformers==2.7.0 --no-deps
chromadb==0.5.3

# ── Document loaders ─────────────────────────────────────────────────
pypdf==4.3.1

# ── API ──────────────────────────────────────────────────────────────
fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-multipart==0.0.9

# ── Dashboard ────────────────────────────────────────────────────────
streamlit==1.36.0
plotly==5.22.0

# ── Utilities ────────────────────────────────────────────────────────
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.5.0
rich==13.7.1
tqdm==4.66.4

# ── Testing ──────────────────────────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.8